Time: 1 ms (84.67%), Space: 43.2 MB (48.78%) - LeetHub

+++ 1945-sum-of-digits-of-string-after-convert/1945-sum-of-digits-of-string-after-convert.java
@@ -1,5 +1,5 @@
 class Solution {
-    int digitsum(long num){
+    int digitsum(int num){
         int sum=0;
         while(num>0){
             sum+=num%10;
@@ -8,7 +8,7 @@ class Solution {
         return sum;
     }
     public int getLucky(String s, int k) {
-        long num=0;
+        int num=0;
         for(int i=0;i<s.length();i++){
             char ch=s.charAt(i);
             int value=ch-'a'+1;
@@ -18,6 +18,6 @@ class Solution {
         for(int i=1;i<k;i++){
             num=digitsum(num);
         }
-        return (int)num;
+        return num;
     }
 }
